Move RiskCalculations test to unit test
