+++ app/src/androidTest/java/com/nsc/covidscore/MainActivityTest.java
@@ -1,5 +1,7 @@
 package com.nsc.covidscore;
 
+import androidx.test.ext.junit.rules.ActivityScenarioRule;
+import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.rule.ActivityTestRule;
 
 import org.junit.Rule;
@@ -14,25 +16,18 @@ import static androidx.test.espresso.assertion.ViewAssertions.matches;
 import static androidx.test.espresso.matcher.ViewMatchers.*;
 import static org.junit.Assert.*;
 
+@RunWith(AndroidJUnit4.class)
 public class MainActivityTest {
 
     @Rule
-    public ActivityTestRule<MainActivity> activityTestRule
-            = new ActivityTestRule<>(MainActivity.class, true, true);
+    public ActivityScenarioRule<MainActivity> activityScenarioRule =
+            new ActivityScenarioRule<>(MainActivity.class);
 
     @Test
-    public void testCalculation() {
-        // King County 10/23/20
-        assertEquals(RiskCalculation.calculateRisk(24914, 2252782, 20), 20.83, 0);
-        // WA State 10/23/20
-        assertEquals(RiskCalculation.calculateRisk(52948, 7614893, 20), 13.63, 0);
-        // US 10/23/20
-        assertEquals(RiskCalculation.calculateRisk(4831614, 328239523, 20), 26.76, 0);
+    public void canSelectLocation() {
+        onView(withId(R.id.state_spinner)).perform())
     }
 
-    @Test
-    public void testCalculationMapping() {
-        Map<Integer, Double> riskMap = RiskCalculation.getRiskCalculationsMap(TestUtils.activeCases, TestUtils.totalPopulation, TestUtils.groupSizes);
-        TestUtils.assertRiskMapMatchesTestData(riskMap);
-    }
+
+
 }
